# src/main/resources/swagger.yaml
openapi: 3.0.3
info:
  title: Sistema de Estoque Distribuido
  description: API para gerenciamento de estoque.
  version: 1.0.0
  contact:
    name: Jobisvaldo Silva
    email: contato@estoque.com
    url: https://sistema-de-estoque.com.br

servers:
  - url: '{protocol}:{url-base}/api/v1'
    variables:
      url-base:
        enum:
          -  //localhost:8080
        default: //localhost:8080
      protocol:
        enum:
          - http
          - https
        default: https
    description: Servidor de Desenvolvimento

paths:
  /products:
    get:
      tags:
        - products
      operationId: getProducts
      summary: Lista todos os produtos
      description: Retorna uma lista paginada de todos os produtos
      parameters:
        - name: page
          in: query
          description: Número da página (começa em 0)
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: Tamanho da página
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Lista de produtos retornada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductPage'
    post:
      operationId: create
      tags:
        - products
      summary: Cria um novo produto
      description: Cria um novo produto no sistema
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
      responses:
        '201':
          description: Produto criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: Dados inválidos fornecidos

  /stores:
    get:
      tags:
        - stores
      operationId: getStores
      summary: Lista todas as lojas
      description: Retorna uma lista de todas as lojas ativas
      responses:
        '200':
          description: Lista de lojas retornada com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StoreResponse'

  /stocks:
    get:
      tags:
        - stocks
      operationId: getStocks
      summary: Lista estoques
      description: Retorna estoques por loja ou produto
      parameters:
        - name: storeId
          in: query
          description: ID da loja
          schema:
            type: string
        - name: productId
          in: query
          description: ID do produto
          schema:
            type: string
      responses:
        '200':
          description: Lista de estoques
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StockResponse'

  /stocks/purchase:
    post:
      tags:
        - stocks
      operationId: simulatePurchase
      summary: Simula uma compra
      description: Simula uma compra reduzindo o estoque disponível
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PurchaseRequest'
      responses:
        '200':
          description: Compra realizada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseResponse'
        '400':
          description: Estoque insuficiente ou dados inválidos

components:

  schemas:
    CreateProductRequest:
      type: object
      properties:
        sku:
          type: string
          example: "SKU-001"
          minLength: 4
          maxLength: 12
        name:
          type: string
          example: "Camiseta Básica"
        description:
          type: string
          example: "Camiseta 100% algodão"
        price:
          type: number
          format: double
          example: 59.90
        quantity:
          type: integer
          example: 100
        category:
          type: string
          example: "Vestuário"

    ProductResponse:
      type: object
      properties:
        id:
          type: string
          example: p-001
        name:
          type: string
          example: "Camiseta Básica"
        description:
          type: string
          example: "Camiseta 100% algodão"
        price:
          type: number
          format: double
          example: 59.90
        category:
          type: string
          example: "Vestuário"
        quantity:
          type: integer
          example: 100

    ProductPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ProductResponse'
        pageable:
          $ref: '#/components/schemas/PageableResponse'

    PageableResponse:
      type: object
      properties:
        pageNumber:
          type: integer
          example: 0
        pageSize:
          type: integer
          example: 20
        totalElements:
          type: integer
          example: 100
        totalPages:
          type: integer
          example: 5
        last:
          type: boolean
          example: false
    StoreResponse:
      type: object
      properties:
        id:
          type: string
          example: "550e8400-e29b-41d4-a716-446655440001"
        storeCode:
          type: string
          example: "STORE001"
        name:
          type: string
          example: "Loja Centro SP"
        address:
          type: string
          example: "Rua Augusta, 123"
        city:
          type: string
          example: "São Paulo"
        state:
          type: string
          example: "SP"
        zipCode:
          type: string
          example: "01305-000"
        phone:
          type: string
          example: "(11) 1234-5678"
        email:
          type: string
          format: email
          example: "centro.sp@example.com"
        active:
          type: boolean
          example: true

    StockResponse:
      type: object
      properties:
        id:
          type: string
          example: "770e8400-e29b-41d4-a716-446655440001"
        productId:
          type: string
          example: "660e8400-e29b-41d4-a716-446655440001"
        productName:
          type: string
          example: "iPhone 15 Pro"
        productSku:
          type: string
          example: "SMARTPHONE001"
        storeId:
          type: string
          example: "550e8400-e29b-41d4-a716-446655440001"
        storeName:
          type: string
          example: "Loja Centro SP"
        storeCode:
          type: string
          example: "STORE001"
        quantity:
          type: integer
          example: 50
        reservedQuantity:
          type: integer
          example: 5
        availableQuantity:
          type: integer
          example: 45
        minimumStock:
          type: integer
          example: 10
        isLowStock:
          type: boolean
          example: false

    PurchaseRequest:
      required:
        - productId
        - storeId
        - quantity
      type: object
      properties:
        productId:
          type: string
          example: "660e8400-e29b-41d4-a716-446655440001"
        storeId:
          type: string
          example: "550e8400-e29b-41d4-a716-446655440001"
        quantity:
          type: integer
          example: 2
          minimum: 1
        customerId:
          type: string
          example: "customer-123"
        orderId:
          type: string
          example: "ORDER-2024-001"

    PurchaseResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Compra realizada com sucesso"
        orderId:
          type: string
          example: "ORDER-2024-001"
        productName:
          type: string
          example: "iPhone 15 Pro"
        storeName:
          type: string
          example: "Loja Centro SP"
        quantityPurchased:
          type: integer
          example: 2
        remainingStock:
          type: integer
          example: 43